west: Add the command pacakge_fw
This the command west pacakge_fw.
The goal of this command is to simplify sharing a prebuilt firmware.
Often, we build a firmware, we share it, but we are not able to rebuild
it if needed because we don't remember the command line we used, the
options we have selected and so. In addition, sometime, we also need
some additional files to use / flash the firmware on the target.
The west package_fw intends to address all this needs.
This adds a first version of the command that currently create a package
with enough information to rebuild it.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>